+++ src/packaging/kpatch/Makefile.upstream
@@ -48,7 +48,6 @@ dm-vdo-objs := \
 	status-codes.o \
 	string-utils.o \
 	sysfs.o \
-	thread-cond-var.o \
 	thread-device.o \
 	thread-registry.o \
 	thread-utils.o \
